==> Figure_3_related/6_JE_K5_mTmG_epi.Rmd <==
---
title: "JE 2nd using K5mTmG tracing -- Specific analysis for epithelium"
author: "Huan"
date: "8/24/2023"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  tidy = TRUE,
  tidy.opts = list(width.cutoff = 120),
  message = TRUE,
  warning = FALSE,
  cache = TRUE
)
```


This note was focused on epithelial population.

# 1 Setup of analysis environment
## 1.1 Package installation 
Seurat (v3), harmony,ggplot2,magrittr and sctransform packages were installed.

## 1.2 Load environment and packages
```{r, load of packages and setup of the environment}
gc()
rm(list=ls())
library(Seurat)
library(SeuratWrappers)
library(sctransform)
library(harmony)
library(data.table)
library(scales)
library(plotly)
library(ggplot2)
library(ggpointdensity)
library(monocle3)
library(magrittr)
library(dplyr)
library(tidyverse)
library(cowplot)
library(clustree)
library(gridExtra)
library(CellChat)
library(scCustomize)

set.seed(12)
theme_set(theme_bw())

JE_K5_epi <- readRDS(file = "./SeuratObject/JE_K5_epi_raw_20230728.rds")


```

## 1.3 Review saved Seurat objects
```{r, review save object}
DimPlot(JE_K5_epi, reduction = "umap", label = T)
DimPlot(JE_K5_epi, reduction = "umap", group.by = "orig.ident")
DimPlot(JE_K5_epi, reduction = "umap", split.by  = "orig.ident")

```
We reperform the dimension reduction and clustering

## 1.4 Rerun harmony

```{r, scale, normalzation and integration}
JE_K5_epi.list <- SplitObject(object = JE_K5_epi, split.by = "orig.ident")
for (i in 1:length(JE_K5_epi.list)) {
    JE_K5_epi.list[[i]] <- SCTransform(JE_K5_epi.list[[i]], vars.to.regress = "percent.mt", verbose = FALSE)
} #I'm not sure if the split and merge is necessary, but it won't cause trouble
```


Following the discussion in `https://github.com/immunogenomics/harmony/issues/41`
```{r, select integration feature for the merged dataset}
JE_K5_epi.features <- SelectIntegrationFeatures(object.list = JE_K5_epi.list, nfeatures = 3000)
JE_K5_epi <- merge(JE_K5_epi.list[[1]],
                   y = JE_K5_epi.list[2:length(JE_K5_epi.list)],  
                   project = "JE_K5_epi", 
                   merge.data = TRUE)
VariableFeatures(JE_K5_epi) <- JE_K5_epi.features
```


PCA using SCT features


```{r,PCA using SCT features}
JE_K5_epi <- RunPCA(object = JE_K5_epi, assay = "SCT", npcs = 50)
```

```{r, integration and batch effect removal using hamony}
system.time({
    JE_K5_epi <- RunHarmony(object = JE_K5_epi, 
                            assay.use = "SCT",
                            reduction = "pca",
                            dims.use = 1:50,
                            group.by.vars = "orig.ident",
                            plot_convergence = TRUE)
})
```


## 1.5 Probing resolution
*You may probe dims and resolution in details*
```{r, dimension reduction}
JE_K5_epi <- RunPCA(object = JE_K5_epi, assay = "SCT", npcs = 30)
JE_K5_epi <- RunUMAP(object = JE_K5_epi, assay = "SCT", reduction = "harmony", dims = 1:50) #You may probe dims in details
Filter(f = function(x) inherits(JE_K5_epi[[x]], "Graph"), names(JE_K5_epi))
JE_K5_epi <- FindNeighbors(object = JE_K5_epi, assay = "SCT", reduction = "harmony", dims = 1:50) #You may probe dims in details
```
```{r, a necessary S4 command}
Filter(f = function(x) inherits(JE_K5_epi[[x]], "Graph"), names(JE_K5_epi))
```


*For details of probing resolution, refer to `https://cloud.tencent.com/developer/article/1825681`. *
Here we will set an intermediate seurat object innheriting all the metadata from the formal object.
```{r, probing resolution parameter}

JE_K5_epi.resolution <- FindClusters(JE_K5_epi, dims=1:50,resolution = seq(from=0,by=.2,length=10))

```

```{r, plot clustree, fig.height=10, fig.width=10}
clustree(JE_K5_epi.resolution)
```
*Choosing resolution is quite arbitrary and no conclusion has been attained before biological validation*
The `JE_K5_epi.resolution` object has many choice for resultion, such as SCT_snn_res.0.8. You may want to plot and check. Here are some examples:
```{r, probeing resolution}
DimPlot(JE_K5_epi.resolution, reduction = "umap", label = T, group.by = "SCT_snn_res.0.2", split.by = "orig.ident")
DimPlot(JE_K5_epi.resolution, reduction = "umap", label = T, group.by = "SCT_snn_res.0.4", split.by = "orig.ident")
DimPlot(JE_K5_epi.resolution, reduction = "umap", label = T, group.by = "SCT_snn_res.0.6", split.by = "orig.ident")
DimPlot(JE_K5_epi.resolution, reduction = "umap", label = T, group.by = "SCT_snn_res.0.8", split.by = "orig.ident")

```


```{r, marker for cluster 0,fig.height=12, fig.width=5}
FeaturePlot(JE_K5_epi.resolution, features = c("Krt14","Krt5",
                                          "Mki67","Lor","Il24", "Odam","GFP"),
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1, split.by = "orig.ident")
```


Here we choose res = 0.6


```{r, cluster with a suitable resolution}
JE_K5_epi <- FindClusters(object = JE_K5_epi, dims=1:50, resolution = 0.6) #You may probe resolution in details
JE_K5_epi <- FindClusters(JE_K5_epi, dims=1:50,resolution = seq(from=0,by=.2,length=10))
```

```{r, test plot,fig.height=5, fig.width=8}
plots <- DimPlot(JE_K5_epi, group.by = c("orig.ident", "seurat_clusters"), combine = FALSE, pt.size = .2)
plots <- lapply(X = plots, FUN = function(x) x + theme(legend.position = "top") + guides(color = guide_legend(nrow = 5, 
    byrow = TRUE, override.aes = list(size = 4))))
CombinePlots(plots)
```

## 1.6 Initial plots 
After the clustering, we are eager to see the topological relationship between different clusters.
```{r, initial plots}
#group_by_cluster
plot1 = DimPlot(JE_K5_epi, reduction = "umap", label = T)
# group_by_sample
plot2 = DimPlot(JE_K5_epi, reduction = "umap", group.by = "orig.ident")
plot3 = DimPlot(JE_K5_epi, reduction = "umap", split.by  = "orig.ident")

# combine
plotc <- plot1 + plot2
ggsave("./OutputFigure/JE_K5_epi_umap.png", plot = plotc, width = 10, height = 5)
plot3
ggsave("./OutputFigure/JE_K5_epi_umap_split.png", plot = plot3, width = 15, height = 5) # You may set the parameters, especially the "width". Usually each plot will take about 5.5 unit.

DimPlot(JE_K5_epi, reduction = "umap", label = T)
DimPlot(JE_K5_epi, reduction = "umap", group.by = "orig.ident")

```
## 1.7 Save object
```{r, output integrated data}
saveRDS(JE_K5_epi, file = "./SeuratObject/JE_K5_epi_harmony_20230824.rds")
```




# 2 Initial marker genes check
## 2.0 Load package and datasets
```{r,resume load of packages and setup of the environment}
gc()
rm(list=ls())
library(Seurat)
library(SeuratWrappers)
library(sctransform)
library(harmony)
library(data.table)
library(scales)
library(plotly)
library(ggplot2)
library(ggpointdensity)
library(monocle3)
library(magrittr)
library(dplyr)
library(tidyverse)
library(cowplot)
library(clustree)
library(gridExtra)
library(CellChat)
library(scCustomize)
set.seed(12)
theme_set(theme_bw())

JE_K5_epi <- readRDS(file = "./SeuratObject/JE_K5_epi_harmony_20230824.rds")


```


## 2.1 Cell cycle score
```{r, cell cycle score}
# A list of cell cycle markers, from Tirosh et al, 2015, is loaded with Seurat.  We can
# segregate this list into markers of G2/M phase and markers of S phase
s.genes <- cc.genes$s.genes
g2m.genes <- cc.genes$g2m.genes
JE_K5_epi <- CellCycleScoring(JE_K5_epi, s.features = s.genes, g2m.features = g2m.genes, set.ident = TRUE)
```

```{r,Visualize the distribution of cell cycle markers across}
# Visualize the distribution of cell cycle markers across
RidgePlot(JE_K5_epi, features = c("Pcna", "Top2a", "Mcm6", "Mki67"), ncol = 2)
ggsave("./OutputFigure/JE_K5_epi_cell_cycle_RidgePlot.pdf", device = "pdf", width = 25, height = 20, units = "cm")
DimPlot(JE_K5_epi, reduction = "umap", label = T)
ggsave("./OutputFigure/JE_K5_epi_cell_cycle_umap_label.pdf", device = "pdf", width = 18, height = 15, units = "cm")
DimPlot(JE_K5_epi, reduction = "umap", group.by = "orig.ident")
ggsave("./OutputFigure/JE_K5_epi_cell_cycle_umap_group.pdf", device = "pdf", width = 18, height = 15, units = "cm")

DimPlot(JE_K5_epi, reduction = "umap", split.by  = "orig.ident")
ggsave("./OutputFigure/JE_K5_epi_cell_cycle_umap_split.pdf", device = "pdf", width = 20, height = 15, units = "cm")

```
Always remember that if you assign a new ident set, double check the active ident (of course currently it is "Phase"). Now reset the active idents.
```{r, set the active ident to res=0.6}
Idents(JE_K5_epi) <- c("SCT_snn_res.0.6")
```

Feature plot for cell cycle genes
```{r, cell cycle genes feature plot split, fig.height=40, fig.width=20}
FeaturePlot(JE_K5_epi, features = c("Pcna","Top2a",
                                          "Mcm6","Mki67"),
            cols = c("#ddf4f5", "#fa3c84"),
            ncol=2, pt.size = 0.1,split.by = "orig.ident")
ggsave("./OutputFigure/JE_K5_epi_cell_cycle_feature_split.pdf", device = "pdf", width = 20, height = 40,limitsize = FALSE)

```

```{r, cell cycle genes feature plot, fig.height=5, fig.width=5}
FeaturePlot(JE_K5_epi, features = c("Pcna","Top2a",
                                          "Mcm6","Mki67"),
            cols = c("#ddf4f5", "#fa3c84"),
            ncol=2, pt.size = 0.1)
ggsave("./OutputFigure/JE_K5_epi_cell_cycle_feature.pdf", device = "pdf", width = 15, height = 15, units = "cm")

```

## 2.4 Stem cell markers
```{r, stem cell genes feature plot, fig.height=5, fig.width=5}
FeaturePlot(JE_K5_epi, features = c("Lrig1","Bmi1","Axin2","Gli1"),
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.3)
#ggsave("./OutputFigure/suture_mes_stem_feature.pdf", device = "pdf", width = 15, height = 15, units = "cm")

```

Refer to `https://www.jianshu.com/p/ea4ec6445c62`
```{r, stem cell genes feature plot split, fig.height=20, fig.width=10}
FeaturePlot(JE_K5_epi, features = c("Krt14","Krt5",
                                          "Krt17","Epcam"),
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1,split.by = "orig.ident")


```
```{r, vlplot for stem cell markers,fig.height=6, fig.width=10}
VlnPlot(object = JE_K5_epi, features =  c("Krt14","Krt5",
                                          "Odam","Lor"),group.by = "seurat_clusters")
#ggsave("./OutputFigure/suture_mes_stem_Markers_violinplot.pdf", device = "pdf", width = 20, height = 15, units = "cm")
```



# 3 Marker genes identification
## 3.1 Marker genes identification
Identification of marker genes
refer to `https://github.com/satijalab/seurat/issues/2115` Perform DE using RNA rather than SCT

```{r, generally find marker genes}
Idents(JE_K5_epi) <- c("SCT_snn_res.0.6")

DefaultAssay(JE_K5_epi) <- "RNA"
JE_K5_epi <- NormalizeData(JE_K5_epi)
all.genes <- rownames(JE_K5_epi)
JE_K5_epi <- ScaleData(JE_K5_epi, features = all.genes)
JE_K5_epi_markers <- FindAllMarkers(object = JE_K5_epi, only.pos = TRUE, 
                               min.pct = 0.25, thresh.use = 0.25) #identify positive marker genes 
top10 <- JE_K5_epi_markers %>% group_by(cluster) %>% top_n(n = 10, wt = avg_log2FC) # You may change n=10 to other number to get top_n genes as you wish.
top6_markers <- Extract_Top_Markers(marker_dataframe = JE_K5_epi_markers, num_genes = 6, named_vector = FALSE,
    make_unique = TRUE)
```

```{r, output tables for marker genes}
write.csv(JE_K5_epi_markers, file = "./OutputTable/JE_K5_epi_markers_20230724.csv", row.names = FALSE)
write.csv(top10, file = "./OutputTable/JE_K5_epi_top10_markers_20230724.csv", row.names = FALSE)


```

Plot heatmap for marker genes
```{r, heatmapplot, fig.width=10, fig.height=18 }

DoHeatmap(JE_K5_epi, features = top10$gene)
ggsave("./OutputFigure/JE_K5_epi_markers_top10_20230724.pdf", device = "pdf",#adjust filename accordingly
       width = 30, #adjust according to the display of output
       height = 40, #adjust according to the display of output
       units = "cm")

```

```{r,prepare avg expression}
gene_cell_exp <- AverageExpression(JE_K5_epi,
                                   features = top10$gene,
                                   group.by = 'seurat_clusters',
                                   slot = 'data') 
gene_cell_exp <- as.data.frame(gene_cell_exp$RNA)
```


```{r, dotplot for top6 marker genes,fig.height=10, fig.width=10}

Clustered_DotPlot(seurat_object = JE_K5_epi, features = top6_markers)
```

## 3.2 Feature plot for marker genes
## 3.2.1 Top 6 genes in each cluster
```{r,generate gene list for top6 genes}
top6 <- JE_K5_epi_markers %>% group_by(cluster) %>% top_n(n = 6, wt = avg_log2FC) # You may change n=10 to other number to get top_n genes as you wish.

```

```{r, marker for cluster mito percent,fig.height=15, fig.width=15}
for (i in unique(top6$cluster)) {
  genes_to_plot <- top6$gene[top6$cluster== i]
  genes_to_plot <- genes_to_plot[genes_to_plot %in% rownames(JE_K5_epi)] 
  
  if (length(genes_to_plot) > 0) {
    print(
      FeaturePlot(JE_K5_epi, 
                  features = genes_to_plot,
                  cols = c("grey85", brewer.pal(9,"YlOrRd")),
                  ncol=2, 
                  pt.size = 0.1, label=T)
    )
  } else {
    print(paste("No genes to plot for cluster", i))
  }
}

```


# 4 MAGIC manipulation
```{r load of packages and setup of the environment 1}
gc()
rm(list=ls())
library(reticulate)
use_condaenv("MAGIC", required = TRUE)
library(Seurat)
library(SeuratWrappers)
library(sctransform)
library(harmony)
library(data.table)
library(scales)
library(plotly)
library(ggplot2)
library(ggpointdensity)
library(monocle3)
library(magrittr)
library(dplyr)
library(tidyverse)
library(cowplot)
library(clustree)
library(gridExtra)
library(phateR)
library(viridis)
library(readr)
library(Rmagic)



set.seed(12)
theme_set(theme_bw())

JE_K5_epi <- readRDS(file = "./SeuratObject/JE_K5_epi_harmony_20230824.rds")

```

First test python environment
```{r}
library(reticulate)
use_condaenv("MAGIC", required = TRUE)
```

```{r}
py_config()
```

```{r}
magic <- import("magic")
```


Generate expression matrix

```{r}

JE_K5_epi_RNA<-GetAssayData(object = JE_K5_epi[["RNA"]], slot = "count")
JE_K5_epi_RNA_table<-as.data.frame(JE_K5_epi_RNA)

head(JE_K5_epi_RNA_table)
```
Transpose matrix
```{r}
JE_K5_epi_RNA_table_t <- as.matrix(t(JE_K5_epi_RNA_table))

```

```{r}
row.names(JE_K5_epi_RNA_table_t) <-NULL
JE_K5_epi_RNA_table_t[1:5,1:10]
```
Filter genes expression matrix
```{r}
# keep genes expressed in at least 10 cells
keep_cols <- colSums(JE_K5_epi_RNA_table_t > 0) > 10
JE_K5_epi_RNA_table_t <- JE_K5_epi_RNA_table_t[,keep_cols]
# look at the distribution of library sizes
ggplot() +
  geom_histogram(aes(x=rowSums(JE_K5_epi_RNA_table_t)), bins=50) +
  geom_vline(xintercept = 2500, color='red')
```
From the above plot, we can set umi threshold as 2500
```{r}
# keep cells with at least 1000 UMIs
keep_rows <- rowSums(JE_K5_epi_RNA_table_t) > 2500
JE_K5_epi_RNA_table_t <- JE_K5_epi_RNA_table_t[keep_rows,]
```

Normalizing data

We should library size normalize and transform the data prior to MAGIC. Many people use a log transform, which requires adding a “pseudocount” to avoid log(0). We square root instead, which has a similar form but doesn’t suffer from instabilities at zero.
```{r}
JE_K5_epi_RNA_table_t <- library.size.normalize(JE_K5_epi_RNA_table_t)
JE_K5_epi_RNA_table_t <- sqrt(JE_K5_epi_RNA_table_t)
```






## 4.1 Further MAGIC for all genes
Using MAGIC for downstream analysis


```{r}
JE_K5_epi_MAGIC <- magic(JE_K5_epi_RNA_table_t, genes=c("Krt5", "Krt14", "Epcam"))
```

```{r}
ggplot(JE_K5_epi_MAGIC) +
  geom_point(aes(Krt5, Krt14, color=Epcam)) +
  scale_color_viridis(option="B")
```
We can look at the entire smoothed matrix with genes='all_genes', passing the original result to the argument init to avoid recomputing intermediate steps. Note that this matrix may be large and could take up a lot of memory.

```{r}
DefaultAssay(JE_K5_epi) <- "SCT"
JE_K5_epi_MAGIC <- magic(JE_K5_epi, genes="all_genes",
                     t=12, init=JE_K5_epi_MAGIC)
#"Added MAGIC output to MAGIC_SCT. To use it, pass assay='MAGIC_SCT' to downstream methods or set DefaultAssay(seurat_object) <- 'MAGIC_SCT'."
```



```{r}
JE_K5_epi_MAGIC_RNA<-GetAssayData(object = JE_K5_epi_MAGIC[["MAGIC_SCT"]], slot = "data")
JE_K5_epi_MAGIC_RNA_table<-as.data.frame(JE_K5_epi_MAGIC_RNA)

head(JE_K5_epi_MAGIC_RNA_table)
```

This is a matrix for further correlation analysis of any genes of interest.

```{r}
JE_K5_epi_MAGIC_RNA_table_t <- t(JE_K5_epi_MAGIC_RNA_table)
JE_K5_epi_MAGIC_RNA_table_t[1:5,1:10]
```

Test if the correlation between Krt14 and Krt5 correlates with Krt17
```{r}
JE_K5_epi_MAGIC_RNA_table_t <- as.data.frame(JE_K5_epi_MAGIC_RNA_table_t)
ggplot(JE_K5_epi_MAGIC_RNA_table_t) +
  geom_point(aes(JE_K5_epi_MAGIC_RNA_table_t$Krt5, JE_K5_epi_MAGIC_RNA_table_t$Krt14, color=JE_K5_epi_MAGIC_RNA_table_t$Krt17)) +
  scale_color_viridis(option="B")
```


```{r, phate dimension reduction}
JE_K5_epi_MAGIC_PHATE <- phate(JE_K5_epi_MAGIC_RNA_table_t, knn=4, decay=100, t=20)
```

```{r}
phate_cord <- as.data.frame(JE_K5_epi_MAGIC_PHATE[["embedding"]])
head(phate_cord)
phate_cord$cellID <- rownames(phate_cord)
head(phate_cord)

```
```{r}
krt17_exp <- as.data.frame(JE_K5_epi_MAGIC_RNA_table_t["Krt17"])
krt17_exp$cellID <- rownames(krt17_exp)
```

```{r}
krt17_cord_exp <- merge(phate_cord, krt17_exp, by=c("cellID"))
```

```{r}
ggplot(krt17_cord_exp) +
  geom_point(aes(x=krt17_cord_exp$PHATE1, y=krt17_cord_exp$PHATE2, color=krt17_cord_exp$Krt17)) +
  scale_color_viridis(option="B") +
  labs(color="Krt17")
```


```{r}
gene_exp <- as.data.frame(JE_K5_epi_MAGIC_RNA_table_t)
gene_exp$cellID <- rownames(gene_exp)
gene_cord_exp <- merge(phate_cord, gene_exp, by=c("cellID"))
```

```{r}
ggplot(gene_cord_exp) +
  geom_point(aes(x=gene_cord_exp$PHATE1, y=gene_cord_exp$PHATE2, color=gene_cord_exp$Krt14)) +
  scale_color_viridis(option="B") +
  labs(color="Krt14")
```

```{r}
ggplot(gene_cord_exp) +
  geom_point(aes(x=gene_cord_exp$PHATE1, y=gene_cord_exp$PHATE2, color=gene_cord_exp$Krt5)) +
  scale_color_viridis(option="B") +
  labs(color="Krt5")
```

```{r}
saveRDS(JE_K5_epi_MAGIC_PHATE, file = "./SeuratObject/JE_K5_epi_MAGIC_PHATE_harmony_20230728.rds")
save(gene_cord_exp,JE_K5_epi_MAGIC_PHATE, file = "./SeuratObject/JE_K5_epi_MAGIC_PHATE_exp.rda" )
```

## 4.2 Standard Seurat for MAGIC object
*You may probe dims and resolution in details*
```{r, dimension reduction}
JE_K5_epi_MAGIC <- RunUMAP(object = JE_K5_epi_MAGIC, assay = "MAGIC_SCT", reduction = "harmony", dims = 1:50) #You may probe dims in details
Filter(f = function(x) inherits(JE_K5_epi_MAGIC[[x]], "Graph"), names(JE_K5_epi_MAGIC))
JE_K5_epi_MAGIC <- FindNeighbors(object = JE_K5_epi_MAGIC, assay = "MAGIC_SCT", reduction = "harmony", dims = 1:50) #You may probe dims in details
```
```{r}
Filter(f = function(x) inherits(JE_K5_epi_MAGIC[[x]], "Graph"), names(JE_K5_epi_MAGIC))
```



*For details of probing resolution, refer to `https://cloud.tencent.com/developer/article/1825681`. *
Here we will set an intermediate seurat object innheriting all the metadata from the formal object.
```{r, probing resolution parameter}

JE_K5_epi_MAGIC.resolution <- FindClusters(JE_K5_epi_MAGIC, dims=1:50,resolution = seq(from=0,by=.2,length=10))

```

```{r, fig.height=10, fig.width=10}
clustree(JE_K5_epi_MAGIC.resolution)
```
*Choosing resolution is quite arbitrary and no conclusion has been attained before biological validation*
The `suture.resolution` object has many choice for resultion, such as SCT_snn_res.0.6. You may want to plot and check. Here are some examples:
```{r, probeing resolution}
DimPlot(JE_K5_epi_MAGIC.resolution, reduction = "umap", label = T, group.by = "SCT_snn_res.0.2")
DimPlot(JE_K5_epi_MAGIC.resolution, reduction = "umap", label = T, group.by = "SCT_snn_res.0.4")
DimPlot(JE_K5_epi_MAGIC.resolution, reduction = "umap", label = T, group.by = "SCT_snn_res.0.6")
DimPlot(JE_K5_epi_MAGIC.resolution, reduction = "umap", label = T, group.by = "SCT_snn_res.0.8")

```

Here we choose res = 0.6


```{r, cluster with a suitable resolution}
JE_K5_epi_MAGIC <- FindClusters(object = JE_K5_epi_MAGIC, resolution = 0.6) #You may probe resolution in details
```

```{r, test plot,fig.height=5, fig.width=8}
plots <- DimPlot(JE_K5_epi_MAGIC, group.by = c("orig.ident", "seurat_clusters"), combine = FALSE, pt.size = .2)
plots <- lapply(X = plots, FUN = function(x) x + theme(legend.position = "top") + guides(color = guide_legend(nrow = 5, 
    byrow = TRUE, override.aes = list(size = 4))))
CombinePlots(plots)
```

## 4.3 Marker genes identification after MAGIC
```{r}
Idents(JE_K5_epi_MAGIC) <- c("SCT_snn_res.0.6")

DefaultAssay(JE_K5_epi_MAGIC) <- "RNA"
JE_K5_epi_MAGIC <- NormalizeData(JE_K5_epi_MAGIC)
all.genes <- rownames(JE_K5_epi_MAGIC)
JE_K5_epi_MAGIC <- ScaleData(JE_K5_epi_MAGIC, features = all.genes)
JE_K5_epi_MAGIC_markers <- FindAllMarkers(object = JE_K5_epi_MAGIC, only.pos = TRUE, 
                               min.pct = 0.25, thresh.use = 0.25) #identify positive marker genes 
top10 <- JE_K5_epi_MAGIC_markers %>% group_by(cluster) %>% top_n(n = 10, wt = avg_log2FC) # You may change n=10 to other number to get top_n genes as you wish.
top6_markers <- Extract_Top_Markers(marker_dataframe = JE_K5_epi_MAGIC_markers, num_genes = 6, named_vector = FALSE,
    make_unique = TRUE)
```

```{r, output tables for marker genes}
write.csv(JE_K5_epi_MAGIC_markers, file = "./OutputTable/JE_K5_epi_MAGIC_0.6_markers_20230724.csv", row.names = FALSE)
write.csv(top10, file = "./OutputTable/JE_K5_epi_MAGIC_0.6_top10_markers_20230724.csv", row.names = FALSE)


```

## 4.3.2 Marker genes dotplot cluster
```{r, dotplot for top6 marker genes,fig.height=15, fig.width=16}

Clustered_DotPlot(seurat_object = JE_K5_epi_MAGIC, features = top6_markers)
```

```{r, dotplot for top6 marker genes,fig.height=16, fig.width=10}

Clustered_DotPlot(seurat_object = JE_K5_epi_MAGIC, features = top6_markers, k = 12, exp_color_min = -1, exp_color_max = 2,colors_use_exp = PurpleAndYellow())


```


## 4.3.3 Marker genes feature plot
```{r}
top6 <- JE_K5_epi_MAGIC_markers %>% group_by(cluster) %>% top_n(n = 6, wt = avg_log2FC) # You may change n=10 to other number to get top_n genes as you wish.

```

```{r, top 6 gene plot before MAGIC,fig.height=15, fig.width=15}
DefaultAssay(JE_K5_epi_MAGIC) <- "RNA"
for (i in unique(top6$cluster)) {
  genes_to_plot <- top6$gene[top6$cluster== i]
  genes_to_plot <- genes_to_plot[genes_to_plot %in% rownames(JE_K5_epi_MAGIC)] 
  
  if (length(genes_to_plot) > 0) {
    print(
      FeaturePlot(JE_K5_epi_MAGIC, 
                  features = genes_to_plot,
                  cols = c("grey85", brewer.pal(9,"YlOrRd")),
                  ncol=2, 
                  pt.size = 0.1,
                  min.cutoff = 0.1,
                  max.cutoff =5)
    )
  } else {
    print(paste("No genes to plot for cluster", i))
  }
}
```

```{r, top6 gene plot after MAGIC,fig.height=15, fig.width=15}
DefaultAssay(JE_K5_epi_MAGIC) <- "MAGIC_SCT"
for (i in unique(top6$cluster)) {
  genes_to_plot <- top6$gene[top6$cluster== i]
  genes_to_plot <- genes_to_plot[genes_to_plot %in% rownames(JE_K5_epi_MAGIC)] 
  
  if (length(genes_to_plot) > 0) {
    print(
      FeaturePlot(JE_K5_epi_MAGIC, 
                  features = genes_to_plot,
                  label = T, 
                  cols = c("grey85", brewer.pal(9,"YlOrRd")),
                  ncol=2, 
                  pt.size = 0.1,
                  min.cutoff = 0.1,
                  max.cutoff =5)
    )
  } else {
    print(paste("No genes to plot for cluster", i))
  }
}
```

```{r}
FeaturePlot(JE_K5_epi_MAGIC, 
                  features = genes_to_plot,
                  cols = c("grey85", brewer.pal(9,"YlOrRd")),
                  ncol=2, 
                  pt.size = 0.1,
                  min.cutoff = 0.1,
                  max.cutoff =5)
```



## 4.4 Initial plots 
After the clustering, we are eager to see the topological relationship between different clusters.
```{r, initial plots}
#group_by_cluster
plot1 = DimPlot(JE_K5_epi_MAGIC, reduction = "umap", label = T)
# group_by_sample
plot2 = DimPlot(JE_K5_epi_MAGIC, reduction = "umap", group.by = "orig.ident")
plot3 = DimPlot(JE_K5_epi_MAGIC, reduction = "umap", split.by  = "orig.ident")
# combine
plotc <- plot1 + plot2
ggsave("./OutputFigure/JE_K5_epi_MAGIC_umap.png", plot = plotc, width = 10, height = 5)
plot3
ggsave("./OutputFigure/JE_K5_epi_MAGIC_umap_split.png", plot = plot3, width = 20, height = 5) # You may set the parameters, especially the "width". Usually each plot will take about 5.5 unit.

DimPlot(JE_K5_epi_MAGIC, reduction = "umap", label = T)
DimPlot(JE_K5_epi_MAGIC, reduction = "umap", group.by = "orig.ident")

```

```{r, marker for cluster mito percent,fig.height=2, fig.width=3}
library(RColorBrewer)
DefaultAssay(JE_K5_epi_MAGIC) <- "MAGIC_SCT"
FeaturePlot(JE_K5_epi_MAGIC, features = c("percent.mt"),
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1)
```

```{r, marker for cluster mito percent,fig.height=2, fig.width=3}
DefaultAssay(JE_K5_epi_MAGIC) <- "MAGIC_SCT"
FeaturePlot(JE_K5_epi_MAGIC, features = c("Il24"),
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1)
```

```{r, marker for cluster mito percent,fig.height=2, fig.width=3}
DefaultAssay(JE_K5_epi_MAGIC) <- "MAGIC_SCT"
FeaturePlot(JE_K5_epi_MAGIC, features = c("Krt5"),
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1)
```
```{r, marker for cluster mito percent,fig.height=3, fig.width=15}
FeaturePlot(JE_K5_epi_MAGIC, features = c("Odam"),
            split.by="orig.ident",
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1)
```
## 4.4.1 Cell cycle score and feature plot
```{r, cell cycle score}
# A list of cell cycle markers, from Tirosh et al, 2015, is loaded with Seurat.  We can
# segregate this list into markers of G2/M phase and markers of S phase
s.genes <- cc.genes$s.genes
g2m.genes <- cc.genes$g2m.genes
JE_K5_epi_MAGIC <- CellCycleScoring(JE_K5_epi_MAGIC, s.features = s.genes, g2m.features = g2m.genes, set.ident = TRUE)
```

```{r,Visualize the distribution of cell cycle markers across}
# Visualize the distribution of cell cycle markers across
RidgePlot(JE_K5_epi_MAGIC, features = c("Pcna", "Top2a", "Mcm6", "Mki67"), ncol = 2)
ggsave("./OutputFigure/JE_K5_epi_MAGIC_cell_cycle_RidgePlot.pdf", device = "pdf", width = 25, height = 20, units = "cm")
DimPlot(JE_K5_epi_MAGIC, reduction = "umap", label = T)
ggsave("./OutputFigure/JE_K5_epi_MAGIC_cell_cycle_umap_label.pdf", device = "pdf", width = 18, height = 15, units = "cm")
DimPlot(JE_K5_epi_MAGIC, reduction = "umap", group.by = "orig.ident")
ggsave("./OutputFigure/JE_K5_epi_MAGIC_cell_cycle_umap_group.pdf", device = "pdf", width = 18, height = 15, units = "cm")

DimPlot(JE_K5_epi_MAGIC, reduction = "umap", split.by  = "orig.ident")
ggsave("./OutputFigure/JE_K5_epi_MAGIC_cell_cycle_umap_split.pdf", device = "pdf", width = 20, height = 15, units = "cm")

```
Always remember that if you assign a new ident set, double check the active ident (of course currently it is "Phase"). Now reset the active idents.
```{r, set the active ident to res=0.6}
DefaultAssay(JE_K5_epi_MAGIC) <- "MAGIC_SCT"
Idents(JE_K5_epi_MAGIC) <- c("SCT_snn_res.0.6")
```

Feature plot for cell cycle genes
```{r, cell cycle genes feature plot split, fig.height=30, fig.width=20}
FeaturePlot(JE_K5_epi_MAGIC, features = c("Pcna","Top2a",
                                          "Mcm6","Mki67"),
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1,split.by = "orig.ident")
ggsave("./OutputFigure/JE_K5_epi_MAGIC_cell_cycle_feature_split.pdf", device = "pdf", width = 20, height = 40,limitsize = FALSE)

```

```{r, cell cycle genes feature plot, fig.height=5, fig.width=5}
FeaturePlot(JE_K5_epi_MAGIC, features =c("Pcna","Top2a",
                                          "Mcm6","Mki67"),
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.3)

ggsave("./OutputFigure/JE_K5_epi_MAGIC_cell_cycle_feature.pdf", device = "pdf", width = 15, height = 15, units = "cm")

```

## 4.4.2 krt gene markers
```{r, stem cell genes feature plot, fig.height=5, fig.width=5}
FeaturePlot(JE_K5_epi_MAGIC, features = c("Lrig1","Bmi1","Axin2","Gli1"),
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.3)
#ggsave("./OutputFigure/JE_K5_epi_MAGIC_stem_feature.pdf", device = "pdf", width = 15, height = 15, units = "cm")

```

Refer to `https://www.jianshu.com/p/ea4ec6445c62`
```{r, krt genes feature plot split, fig.height=20, fig.width=10}
FeaturePlot(JE_K5_epi_MAGIC, features = c("Lrig1","Bmi1","Axin2","Gli1"),
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1,split.by = "orig.ident")


```
```{r, vlplot for stem cell markers,fig.height=6, fig.width=10}
VlnPlot(object = JE_K5_epi_MAGIC, features =  c("Lrig1","Bmi1","Axin2","Gli1"),group.by = "seurat_clusters")
#ggsave("./OutputFigure/suture_mes_stem_Markers_violinplot.pdf", device = "pdf", width = 20, height = 15, units = "cm")
```

## 4.4.3 krt gene markers
```{r, stem cell genes feature plot, fig.height=5, fig.width=5}
FeaturePlot(JE_K5_epi_MAGIC, features = c("Krt14","Krt5",
                                          "Odam","Lor"),
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.3)
#ggsave("./OutputFigure/JE_K5_epi_MAGIC_stem_feature.pdf", device = "pdf", width = 15, height = 15, units = "cm")

```

Refer to `https://www.jianshu.com/p/ea4ec6445c62`
```{r, krt genes feature plot split, fig.height=20, fig.width=10}
FeaturePlot(JE_K5_epi_MAGIC, features = c("Krt14","Krt5",
                                          "Odam","Lor"),
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1,split.by = "orig.ident")


```
```{r, vlplot for stem cell markers,fig.height=6, fig.width=10}
VlnPlot(object = JE_K5_epi_MAGIC, features =  c("Krt14","Krt5",
                                          "Odam","Lor"),group.by = "seurat_clusters")
#ggsave("./OutputFigure/suture_mes_stem_Markers_violinplot.pdf", device = "pdf", width = 20, height = 15, units = "cm")
```

## 4.4.5 GFP gene markers
```{r, stem cell genes feature plot, fig.height=5, fig.width=5}
FeaturePlot(JE_K5_epi_MAGIC, features = c("GFP","tdTomato",
                                          "Odam","Krt5"),
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.3)
#ggsave("./OutputFigure/JE_K5_epi_MAGIC_stem_feature.pdf", device = "pdf", width = 15, height = 15, units = "cm")

```

Refer to `https://www.jianshu.com/p/ea4ec6445c62`
```{r, krt genes feature plot split, fig.height=15, fig.width=10}
FeaturePlot(JE_K5_epi_MAGIC, features = c("GFP","tdTomato",
                                          "Odam","Krt5"),
            min.cutoff=0.2,
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1,split.by = "orig.ident")


```

```{r, krt genes feature plot split, fig.height=4, fig.width=10}
FeaturePlot(JE_K5_epi_MAGIC, features = c("GFP"),
            min.cutoff=0.2,
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1,split.by = "orig.ident")


```
```{r, krt genes feature plot split, fig.height=4, fig.width=10}
FeaturePlot(JE_K5_epi_MAGIC, features = c("tdTomato"),
            min.cutoff=0.4,
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1,split.by = "orig.ident")


```

```{r, vlplot for stem cell markers,fig.height=6, fig.width=10}
VlnPlot(object = JE_K5_epi_MAGIC, features =  c("GFP","tdTomato",
                                          "Odam","Krt5"),group.by = "seurat_clusters")
#ggsave("./OutputFigure/suture_mes_stem_Markers_violinplot.pdf", device = "pdf", width = 20, height = 15, units = "cm")
```
```{r, krt genes feature plot split, fig.height=4, fig.width=10}
DefaultAssay(JE_K5_epi_MAGIC) <- "RNA"
FeaturePlot(JE_K5_epi_MAGIC, features = c("tdTomato"),
            min.cutoff=0.4,
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1,split.by = "orig.ident")
FeaturePlot(JE_K5_epi_MAGIC, features = c("GFP"),
            min.cutoff=0.2,
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1,split.by = "orig.ident")


```
```{r, krt genes feature plot split, fig.height=4, fig.width=10}
DefaultAssay(JE_K5_epi_MAGIC) <- "SCT"
FeaturePlot(JE_K5_epi_MAGIC, features = c("tdTomato"),
            min.cutoff=0.4,
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1,split.by = "orig.ident")
FeaturePlot(JE_K5_epi_MAGIC, features = c("GFP"),
            min.cutoff=0.2,
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1,split.by = "orig.ident")


```


```{r, krt genes feature plot split, fig.height=4, fig.width=10}
DefaultAssay(JE_K5_epi_MAGIC) <- "MAGIC_SCT"
FeaturePlot(JE_K5_epi_MAGIC, features = c("tdTomato"),
            min.cutoff=0.4,
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1,split.by = "orig.ident")
ggsave("./OutputFigure/JE_K5_epi_MAGIC_tdTomato_feature.pdf", device = "pdf", width = 10, height = 4)
FeaturePlot(JE_K5_epi_MAGIC, features = c("GFP"),
            min.cutoff=0.2,
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1,split.by = "orig.ident")
ggsave("./OutputFigure/JE_K5_epi_MAGIC_GFP_feature.pdf", device = "pdf", width = 10, height = 4)
FeaturePlot(JE_K5_epi_MAGIC, features = c("Krt5"),
            min.cutoff=0.2,
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1,split.by = "orig.ident")
ggsave("./OutputFigure/JE_K5_epi_MAGIC_Krt5_feature.pdf", device = "pdf", width = 10, height = 4)
FeaturePlot(JE_K5_epi_MAGIC, features = c("Odam"),
            min.cutoff=0.2,
            cols = c("grey85",brewer.pal(9,"YlOrRd")),
            ncol=2, pt.size = 0.1,split.by = "orig.ident")
ggsave("./OutputFigure/JE_K5_epi_MAGIC_Odam_feature.pdf", device = "pdf", width = 10, height = 4)
```
## 4.5 Save unnamed MAGIC object
```{r, output integrated data}
saveRDS(JE_K5_epi_MAGIC, file = "./SeuratObject/JE_K5_epi_MAGIC_harmony_20230728.rds")
```

```{r}
rm(list=ls())
gc()

```

```{r}
JE_K5_epi_MAGIC <- readRDS("./SeuratObject/JE_K5_epi_MAGIC_harmony_20230728.rds")
```




## 4.5 Rename cluster
```{r,Load packages and datasets 5}
gc()
rm(list=ls())
library(Seurat)
library(SeuratWrappers)
library(data.table)
library(scales)
library(plotly)
library(ggplot2)
library(ggpointdensity)
library(magrittr)
library(velocyto.R)
library(dplyr)
library(tidyverse)
library(cowplot)
library(gridExtra)

set.seed(12)
theme_set(theme_bw())

JE_K5_epi_MAGIC <- readRDS(file = "./SeuratObject/JE_K5_epi_MAGIC_harmony_20230728.rds")

```



```{r, test plot,fig.height=5, fig.width=8}
plots <- DimPlot(JE_K5_epi_MAGIC, group.by = c("orig.ident", "seurat_clusters"), combine = FALSE, pt.size = .2, label=T, repel=T)
plots <- lapply(X = plots, FUN = function(x) x + theme(legend.position = "top") + guides(color = guide_legend(nrow = 5, 
    byrow = TRUE, override.aes = list(size = 4))))
CombinePlots(plots)
```

With "e" stands for epithelium, we renamed the following clusters.
```{r, rename cluster}
new.cluster.ids <- c("Basal", #Cluster 0 
                     "Intermidate", #Cluster 1 
                     "Basal", #Cluster 2 
                     "Keratinocyte", #Cluster 3 
                     "Basal", #Cluster 4 
                     "Cycling", #Cluster 5 
                     "Basal", #Cluster 6 
                     "Intermidate", #Cluster 7 
                     "Intermidate", #Cluster 8 
                     "Intermidate", #Cluster 9 
                     "Basal", #Cluster 10 
                     "JE", #Cluster 11
                     "Basal" #Cluster 12 
                     )
Idents(JE_K5_epi_MAGIC) <- c("SCT_snn_res.0.6")
JE_K5_epi_MAGIC@meta.data$seurat_clusters <-JE_K5_epi_MAGIC@meta.data$SCT_snn_res.0.6
JE_K5_epi_MAGIC@meta.data$celltype<- JE_K5_epi_MAGIC@meta.data$seurat_clusters
levels(JE_K5_epi_MAGIC@meta.data$celltype) <- new.cluster.ids
```

```{r,draft plot for celltype}
DimPlot(JE_K5_epi_MAGIC, reduction = "umap", label = T, group.by = "celltype")
```

```{r, initial plots}
Idents(JE_K5_epi_MAGIC) <- c("celltype")
#group_by_cluster
plot1 = DimPlot(JE_K5_epi_MAGIC, reduction = "umap", label = T)
# group_by_sample
plot2 = DimPlot(JE_K5_epi_MAGIC, reduction = "umap", group.by = "celltype")

# combine
plotc <- plot1 + plot2
ggsave("./OutputFigure/JE_K5_epi_MAGIC_umap_celltype.png", plot = plotc, width = 10, height = 5)


DimPlot(JE_K5_epi_MAGIC, reduction = "umap", label = T)
DimPlot(JE_K5_epi_MAGIC, reduction = "umap", group.by = "celltype")

```

```{r, output integrated data}
saveRDS(JE_K5_epi_MAGIC, file = "./SeuratObject/JE_K5_epi_MAGIC_harmony_renamed.rds")
```
